+++ tools/test-js-with-node
@@ -160,7 +160,6 @@ EXEMPT_FILES = make_set(
         "web/src/settings_exports.js",
         "web/src/settings_invites.js",
         "web/src/settings_linkifiers.js",
-        "web/src/settings_muted_topics.js",
         "web/src/settings_muted_users.js",
         "web/src/settings_notifications.js",
         "web/src/settings_org.js",
@@ -174,6 +173,7 @@ EXEMPT_FILES = make_set(
         "web/src/settings_toggle.js",
         "web/src/settings_ui.js",
         "web/src/settings_user_groups_legacy.js",
+        "web/src/settings_user_topics.js",
         "web/src/settings_users.js",
         "web/src/setup.js",
         "web/src/spectators.js",
