+++ web/src/settings_sections.js
@@ -8,7 +8,6 @@ import * as settings_emoji from "./settings_emoji";
 import * as settings_exports from "./settings_exports";
 import * as settings_invites from "./settings_invites";
 import * as settings_linkifiers from "./settings_linkifiers";
-import * as settings_muted_topics from "./settings_muted_topics";
 import * as settings_muted_users from "./settings_muted_users";
 import * as settings_notifications from "./settings_notifications";
 import * as settings_org from "./settings_org";
@@ -17,6 +16,7 @@ import * as settings_profile_fields from "./settings_profile_fields";
 import * as settings_realm_user_settings_defaults from "./settings_realm_user_settings_defaults";
 import * as settings_streams from "./settings_streams";
 import * as settings_user_groups from "./settings_user_groups_legacy";
+import * as settings_user_topics from "./settings_user_topics";
 import * as settings_users from "./settings_users";
 
 const load_func_dict = new Map(); // group -> function
@@ -60,7 +60,7 @@ export function initialize() {
     load_func_dict.set("your-bots", settings_bots.set_up);
     load_func_dict.set("alert-words", alert_words_ui.set_up_alert_words);
     load_func_dict.set("uploaded-files", attachments_ui.set_up_attachments);
-    load_func_dict.set("topics", settings_muted_topics.set_up);
+    load_func_dict.set("topics", settings_user_topics.set_up);
     load_func_dict.set("muted-users", settings_muted_users.set_up);
 
     // org
@@ -113,7 +113,7 @@ export function reset_sections() {
     settings_profile_fields.reset();
     settings_streams.reset();
     settings_user_groups.reset();
-    settings_muted_topics.reset();
+    settings_user_topics.reset();
     settings_muted_users.reset();
     alert_words_ui.reset();
     // settings_users doesn't need a reset()
